// src/app/generated/model/lowStockItemResponse.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface LowStockItemResponse { 
    /**
     * Identifier of the item
     */
    itemId?: number;
    /**
     * Name of the item
     */
    itemName?: string;
    /**
     * Stock keeping unit code of the item
     */
    sku?: string;
    /**
     * Identifier of the warehouse holding the stock
     */
    warehouseId?: number;
    /**
     * Name of the warehouse holding the stock
     */
    warehouseName?: string;
    /**
     * Current quantity on hand
     */
    quantity?: number;
    /**
     * Quantity threshold below which the item is considered low stock
     */
    reorderThreshold?: number;
}