// src/app/generated/model/stockLevelResponse.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StockLevelResponse { 
    /**
     * Identifier of the item
     */
    itemId?: number;
    /**
     * Name of the item
     */
    itemName?: string;
    /**
     * Stock keeping unit code of the item
     */
    sku?: string;
    /**
     * Identifier of the warehouse holding the stock
     */
    warehouseId?: number;
    /**
     * Name of the warehouse holding the stock
     */
    warehouseName?: string;
    /**
     * Identifier of the storage location within the warehouse, if tracked
     */
    locationId?: number;
    /**
     * Code of the storage location within the warehouse, if tracked
     */
    locationCode?: string;
    /**
     * Current quantity on hand
     */
    quantity?: number;
    /**
     * Quantity threshold below which the item is considered low stock, if configured
     */
    reorderThreshold?: number;
}