// src/app/generated/model/stockTransferRequest.ts
/**
 * BC Construction Services API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StockTransferRequest { 
    /**
     * Identifier of the item being transferred
     */
    itemId: number;
    /**
     * Identifier of the warehouse the stock is being moved from
     */
    fromWarehouseId: number;
    /**
     * Identifier of the storage location being moved from, if tracked
     */
    fromLocationId?: number;
    /**
     * Identifier of the warehouse the stock is being moved to
     */
    toWarehouseId: number;
    /**
     * Identifier of the storage location being moved to, if tracked
     */
    toLocationId?: number;
    /**
     * Quantity of stock to transfer
     */
    quantity: number;
}